fix: respect timerange when loading trades from parquet files
ParquetDataHandler ignored the timerange parameter when loading trades, while FeatherDataHandler already supported Arrow-based filtering. Move _build_arrow_time_filter to IDataHandler and apply the same filtering logic to parquet trade loads.
