+++ freqtrade/data/history/datahandlers/featherdatahandler.py
@@ -112,36 +112,6 @@ class FeatherDataHandler(IDataHandler):
         """
         raise NotImplementedError()
 
-    def _build_arrow_time_filter(self, timerange: TimeRange | None):
-        """
-        Build Arrow predicate filter for timerange filtering.
-        Treats 0 as unbounded (no filter on that side).
-        :param timerange: TimeRange object with start/stop timestamps
-        :return: Arrow filter expression or None if fully unbounded
-        """
-        if not timerange:
-            return None
-
-        # Treat 0 as unbounded
-        start_set = bool(timerange.startts and timerange.startts > 0)
-        stop_set = bool(timerange.stopts and timerange.stopts > 0)
-
-        if not (start_set or stop_set):
-            return None
-
-        ts_field = dataset.field("timestamp")
-        exprs = []
-
-        if start_set:
-            exprs.append(ts_field >= timerange.startts)
-        if stop_set:
-            exprs.append(ts_field <= timerange.stopts)
-
-        if len(exprs) == 1:
-            return exprs[0]
-        else:
-            return exprs[0] & exprs[1]
-
     def _trades_load(
         self, pair: str, trading_mode: TradingMode, timerange: TimeRange | None = None
     ) -> DataFrame:
